# x-api/src/docs/swagger.yaml
basePath: /api/v1
definitions:
  apientities.CeleryStatus:
    properties:
      general:
        example: SUCCESS
        format: string
        type: string
      ml:
        example: SUCCESS
        format: string
        type: string
      upload:
        example: SUCCESS
        format: string
        type: string
    type: object
  apientities.CeleryStatusResult:
    properties:
      status:
        $ref: '#/definitions/apientities.CeleryStatus'
      task_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: string
        type: string
      times:
        $ref: '#/definitions/apientities.CeleryTimes'
    type: object
  apientities.CeleryTimes:
    properties:
      end_ml:
        example: "2021-12-29 10:11:22"
        format: string
        type: string
      end_upload:
        example: "2021-12-29 10:11:22"
        format: string
        type: string
      start_ml:
        example: "2021-12-29 10:11:22"
        format: string
        type: string
      start_upload:
        example: "2021-12-29 10:11:22"
        format: string
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: duynguyenngoc@hotmail.com
    name: Duy Nguyen
  description: This is a x-ml api server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger x-api
  version: "1.0"
paths:
  /celery/status/{task_id}:
    get:
      consumes:
      - application/json
      description: show result celery status base on task id
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apientities.CeleryStatusResult'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Celery Status
      tags:
      - Celery
swagger: "2.0"